+++ src/lightning_habana/accelerators/__init__.py
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from lightning_habana.fabric.plugins.io.hpu_io import HPUCheckpointIO
+from lightning_habana.accelerators.fabric_accel import HPUAccelerator
 
-__all__ = ["HPUCheckpointIO"]
+__all__ = ["HPUAccelerator"]
