+++ src/lightning_habana/strategies/__init__.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from lightning_habana.strategies.parallel import HPUParallelStrategy
-from lightning_habana.strategies.single import SingleHPUStrategy
+from lightning_habana.strategies.pl_parallel import HPUParallelStrategy
+from lightning_habana.strategies.pl_single import SingleHPUStrategy
 
 __all__ = ["HPUParallelStrategy", "SingleHPUStrategy"]
