+++ src/lightning_habana/strategies/fabric_parallel.py
@@ -29,7 +29,7 @@ from lightning.fabric.utilities.types import Optimizable
 from torch import Tensor
 from torch.nn import Module
 
-from lightning_habana.fabric.accelerator.hpu import HPUAccelerator
+from lightning_habana.fabric.accelerator.fabric_accel import HPUAccelerator
 from lightning_habana.fabric.plugins.precision import HPUPrecision
 from lightning_habana.fabric.utils.imports import _HPU_AVAILABLE, _TORCH_LESSER_EQUAL_1_13_1
 
@@ -40,7 +40,7 @@ if _HPU_AVAILABLE:
 log = logging.getLogger(__name__)
 
 
-class HPUParallelStrategy(DDPStrategy):
+class FabricHPUParallelStrategy(DDPStrategy):
     """Strategy for distributed training on multiple HPU devices."""
 
     strategy_name = "parallel_hpu"
